backend/requirements.txt
Pillow
python-dotenv
drf-extra-fields
reportlab
fontTools
rinoh-typeface-dejavuserif
asgiref
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
django==3.2.5
djangorestframework==3.11.0
idna==2.9
importlib-metadata==1.6.0
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.7
pytest-django==3.9.0
pytest==5.4.1
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
wcwidth==0.1.9
zipp==3.1.0
django-filter==2.4.0
djangorestframework-simplejwt==4.6.0
gunicorn==20.1.0
djoser
psycopg2-binary==2.8.5
drf-spectacular